+++ src/main/java/codyhuh/worldofwonder/WorldOfWonder.java
@@ -10,14 +10,14 @@ import codyhuh.worldofwonder.common.compat.WonderVanillaCompat;
 import codyhuh.worldofwonder.common.util.RegistryHelper;
 import codyhuh.worldofwonder.common.world.WonderFeatureAndBiomeGenerator;
 import codyhuh.worldofwonder.common.world.biome.DandelionFieldsRegionProvider;
-import codyhuh.worldofwonder.init.WonderBlocks;
-import codyhuh.worldofwonder.init.WonderEntities;
-import codyhuh.worldofwonder.init.WonderFoliagePlacers;
-import codyhuh.worldofwonder.init.WonderItems;
-import codyhuh.worldofwonder.init.WonderSounds;
-import codyhuh.worldofwonder.init.WonderTabs;
-import codyhuh.worldofwonder.init.WonderTileEntities;
-import codyhuh.worldofwonder.init.WonderWoodTypes;
+import codyhuh.worldofwonder.core.WonderBlocks;
+import codyhuh.worldofwonder.core.WonderEntities;
+import codyhuh.worldofwonder.core.WonderFoliagePlacers;
+import codyhuh.worldofwonder.core.WonderItems;
+import codyhuh.worldofwonder.core.WonderSounds;
+import codyhuh.worldofwonder.core.WonderTabs;
+import codyhuh.worldofwonder.core.WonderTileEntities;
+import codyhuh.worldofwonder.core.WonderWoodTypes;
 import net.minecraft.core.HolderLookup;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.data.DataGenerator;
