+++ src/main/java/codyhuh/worldofwonder/client/ClientEvents.java
@@ -8,9 +8,9 @@ import codyhuh.worldofwonder.client.renderer.entity.model.DandeLionModel;
 import codyhuh.worldofwonder.client.renderer.entity.model.DandelionHatModel;
 import codyhuh.worldofwonder.client.renderer.tileentity.StemChestTileEntityRenderer;
 import codyhuh.worldofwonder.common.compat.WonderQuarkCompat;
-import codyhuh.worldofwonder.init.WonderEntities;
-import codyhuh.worldofwonder.init.WonderTileEntities;
-import codyhuh.worldofwonder.init.WonderWoodTypes;
+import codyhuh.worldofwonder.core.WonderEntities;
+import codyhuh.worldofwonder.core.WonderTileEntities;
+import codyhuh.worldofwonder.core.WonderWoodTypes;
 import net.minecraft.client.model.geom.ModelLayerLocation;
 import net.minecraft.client.model.geom.builders.CubeDeformation;
 import net.minecraft.client.renderer.Sheets;
