+++ src/main/java/codyhuh/worldofwonder/client/WonderClientHandler.java
@@ -5,8 +5,8 @@ import codyhuh.worldofwonder.client.renderer.entity.DandeLionSeedRenderer;
 import codyhuh.worldofwonder.client.renderer.entity.StemBoatRenderer;
 import codyhuh.worldofwonder.client.renderer.tileentity.StemChestTileEntityRenderer;
 import codyhuh.worldofwonder.common.compat.WonderQuarkCompat;
-import codyhuh.worldofwonder.init.WonderBlocks;
-import codyhuh.worldofwonder.init.WonderEntities;
+import codyhuh.worldofwonder.core.WonderBlocks;
+import codyhuh.worldofwonder.core.WonderEntities;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.client.renderer.blockentity.BlockEntityRenderers;
