+++ src/main/java/codyhuh/worldofwonder/common/block/trees/DandelionTree.java
@@ -2,7 +2,7 @@ package codyhuh.worldofwonder.common.block.trees;
 
 import javax.annotation.Nullable;
 
-import codyhuh.worldofwonder.init.WonderFeatures;
+import codyhuh.worldofwonder.core.WonderFeatures;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.block.grower.AbstractTreeGrower;
