+++ src/main/java/codyhuh/worldofwonder/common/events/ItemEvents.java
@@ -3,8 +3,8 @@ package codyhuh.worldofwonder.common.events;
 import codyhuh.worldofwonder.WorldOfWonder;
 import codyhuh.worldofwonder.common.block.trees.DandelionFluffTree;
 import codyhuh.worldofwonder.common.block.trees.DandelionTree;
-import codyhuh.worldofwonder.init.WonderBlocks;
-import codyhuh.worldofwonder.init.WonderItems;
+import codyhuh.worldofwonder.core.WonderBlocks;
+import codyhuh.worldofwonder.core.WonderItems;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.BlockSource;
 import net.minecraft.core.dispenser.DispenseItemBehavior;
