+++ src/main/java/codyhuh/worldofwonder/common/tileentity/WonderHangingSignBlockEntity.java
@@ -1,6 +1,6 @@
 package codyhuh.worldofwonder.common.tileentity;
 
-import codyhuh.worldofwonder.init.WonderTileEntities;
+import codyhuh.worldofwonder.core.WonderTileEntities;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.entity.HangingSignBlockEntity;
