+++ src/main/java/codyhuh/worldofwonder/common/world/biome/DandelionFieldsBiomeDecorator.java
@@ -1,7 +1,7 @@
 package codyhuh.worldofwonder.common.world.biome;
 
-import codyhuh.worldofwonder.init.WonderEntities;
-import codyhuh.worldofwonder.init.WonderPlacedFeatures;
+import codyhuh.worldofwonder.core.WonderEntities;
+import codyhuh.worldofwonder.core.WonderPlacedFeatures;
 import net.minecraft.core.HolderGetter;
 import net.minecraft.data.worldgen.BiomeDefaultFeatures;
 import net.minecraft.data.worldgen.placement.VegetationPlacements;
