+++ src/main/java/codyhuh/worldofwonder/common/world/biome/DandelionFieldsRegionProvider.java
@@ -4,7 +4,7 @@ import java.util.function.Consumer;
 
 import com.mojang.datafixers.util.Pair;
 
-import codyhuh.worldofwonder.init.WonderBiomes;
+import codyhuh.worldofwonder.core.WonderBiomes;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
