+++ src/main/java/codyhuh/worldofwonder/core/WonderSounds.java
@@ -1,4 +1,4 @@
-package codyhuh.worldofwonder.init;
+package codyhuh.worldofwonder.core;
 
 import codyhuh.worldofwonder.WorldOfWonder;
 import net.minecraft.resources.ResourceLocation;
