+++ src/main/java/codyhuh/worldofwonder/core/WonderTabs.java
@@ -1,4 +1,4 @@
-package codyhuh.worldofwonder.init;
+package codyhuh.worldofwonder.core;
 
 import codyhuh.worldofwonder.WorldOfWonder;
 import net.minecraft.core.registries.Registries;
